Add webhook signature verification example coming from a Bullet Train app
First we super-scaffolded a new incoming webhooks controller `rails g super_scaffold:incoming_webhooks BulletTrain`.

Then, the `verify` and `verify_request` methods were taken from the Bullet Train repo to verify the signature in the controller. If a signature is not verified, we don't save a new Webhook record to the DB.